fix(grid): renamed grid demo components, edited properties
