+++ src/components/grid/__tests__/GridUsageExample.spec.js
@@ -1,0 +1,44 @@
+import React from 'react'
+import { render } from 'react-testing-library'
+import { ThemeProvider } from 'react-jss'
+import theme from 'zeppelin-element-library/bundle/themes/theme.json'
+import GridUsageExample from '../GridUsageExample'
+
+describe('GridUsageExample', () => {
+  it('renders correctly', () => {
+    const mockProps = {
+      rows: [
+        {
+          cols: [
+            {
+              xxs: 3,
+              xs: 3,
+              m: 2,
+              l: 3,
+              text: 'xxs=1 xs=3 m=2 l=3',
+            },
+            {
+              xxs: 3,
+              xs: 3,
+              m: 2,
+              l: 3,
+              text: 'xxs=1 xs=3 m=2 l=3',
+            },
+            {
+              xxs: 4,
+              m: 4,
+              l: 6,
+              text: 'xxs=4 m=4 l=6',
+            },
+          ],
+        },
+      ],
+    }
+    const component = render(
+      <ThemeProvider theme={theme}>
+        <GridUsageExample {...mockProps} />
+      </ThemeProvider>
+    )
+    expect(component).toMatchSnapshot()
+  })
+})
